+++ src/main/java/com/eauction/www/auction/models/Result.java
@@ -12,5 +12,5 @@ public class Result {
     private String itemId;
     private String bidder;
 
-    private Integer bidAmount;
+    private Double bidAmount;
 }
